+++ src/main/java/com/LetUsCodeTogether/ats/repo/OverallRankRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface OverallRankRepository extends JpaRepository<OverallRank, Integer> {
-    Optional<OverallRank> findByUserId(Integer userId);
+    Optional<OverallRank> findByUserId(long userId);
 }
